+++ configurations/liquibase/src/main/java/io/micronaut/configuration/dbmigration/liquibase/AbstractLiquibase.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.micronaut.dbmigration.liquibase;
+package io.micronaut.configuration.dbmigration.liquibase;
 
 import io.micronaut.core.util.StringUtils;
 import liquibase.Contexts;
