+++ configurations/liquibase/src/test/groovy/io/micronaut/configuration/dbmigration/liquibase/LiquibaseConfigurationPropertiesEnabledSpec.groovy
@@ -1,9 +1,24 @@
-package io.micronaut.docs.dbmigration.liquibase
+/*
+ * Copyright 2017-2018 original authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.micronaut.configuration.dbmigration.liquibase
 
 import io.micronaut.context.ApplicationContext
 import io.micronaut.context.env.Environment
 import io.micronaut.context.exceptions.NoSuchBeanException
-import io.micronaut.dbmigration.liquibase.LiquibaseConfigurationProperties
 import io.micronaut.inject.qualifiers.Qualifiers
 import spock.lang.Specification
