+++ configurations/liquibase/src/test/groovy/io/micronaut/configuration/dbmigration/liquibase/LiquibaseDisabledSpec.groovy
@@ -14,12 +14,11 @@
  * limitations under the License.
  */
 
-package io.micronaut.docs.dbmigration.liquibase
+package io.micronaut.configuration.dbmigration.liquibase
 
 import io.micronaut.context.ApplicationContext
 import io.micronaut.context.env.Environment
 import io.micronaut.context.exceptions.NoSuchBeanException
-import io.micronaut.dbmigration.liquibase.LiquibaseConfigurationProperties
 import io.micronaut.testutils.YamlAsciidocTagCleaner
 import spock.lang.AutoCleanup
 import spock.lang.Shared
